refactor the use of delete and change name column reducers
