+++ src/components/ColumnHeader.tsx
@@ -1,21 +1,37 @@
 import { useState } from "react"
 import './ColumnHeader.css'
 import { PencilSquareIcon, TrashIcon } from "./Icon"
+import { useDispatch } from "react-redux"
+import { changeColumnName, deleteColumn } from "../redux/columnsSlice"
+import toast from "react-hot-toast"
 
 interface ColumnHeaderProps{
     name: string
     columnId: string
-    changeColumnName: Function,
-    deleteColumn: Function
 }
 
-function ColumnHeader({ name, columnId, changeColumnName, deleteColumn }: ColumnHeaderProps) {
+function ColumnHeader({ name, columnId }: ColumnHeaderProps) {
     const [ columnName, setColumnName ] = useState(name)
     const [ isTheColumnNameChanging, setIsTheColumnNameChanging ] = useState(false)
+    const dispatch = useDispatch();
+
+    const changeTheColumnName = (columnId: string, newColumnName: string) => {
+      dispatch(changeColumnName({ columnId, newColumnName }))
+    }
+  
+    const deleteTheColumn = (columnId: string) => {
+      try {
+        dispatch(deleteColumn(columnId))
+        toast.success('Columna eliminada')
+      }
+      catch(e) {
+        toast.error('Solo hay tres columnas, no se puede eliminar esta columna')
+      }
+    }
 
     const editColumnName = () => {
         if(isTheColumnNameChanging && !!columnName.trim()) {
-            changeColumnName(columnId, columnName)
+            changeTheColumnName(columnId, columnName)
         }
         if(!columnName.trim()) {
             setColumnName(name)
@@ -33,7 +49,7 @@ function ColumnHeader({ name, columnId, changeColumnName, deleteColumn }: Column
             editColumnName()
         }
         if(columnAction == columnActions.DELETE) {
-            deleteColumn(columnId)
+            deleteTheColumn(columnId)
         }
     }
